refactor(runtime): extract runtime settings application layer
